Improves Span Overlay and Trace Detail Layout (#57)
* Adds collapsible sections to span details panel

Implements collapsible sections for the span details panel to improve readability and organization.

Introduces an accordion-like structure for additional span data, events, and process information.

This allows users to focus on relevant information and reduces visual clutter.

* Improves span overlay and trace detail layout.

Moves the close button in the span overlay to be absolute, improving layout.

Compensates for Grafana panel padding in trace detail to prevent content overflow.

* Adds accordion component for details panel

Implements an accordion component to display span details within the panel.
This improves the user experience by allowing users to expand and collapse
sections of the details panel, making it easier to navigate and focus on
specific information. It also extracts process and event details for display.

* Distinguish between span and resource tags

* Ensure resource tag are fetched with the right prefix.

* Adds spacing to the Accordion component

Improves the visual separation of the Accordion component from surrounding elements by adding a margin-top.

---------

Co-authored-by: nojaf <florian.verdonck@outlook.com>